The variable type for response_time got changed from float to double.  This broke the fscanf that was trying to put a value into response_time.
git-svn-id: https://osdldbt.svn.sourceforge.net/svnroot/osdldbt/trunk/dbt1@9 eccdf8fb-1411-0410-9b6e-c06b97aefaee

+++ tools/results.c
@@ -40,7 +40,7 @@ int main(int argc, char *argv[])
 	FILE *plot_ips;
 	time_t current_time;
 	char interaction[3];
-	double response_time;
+	float response_time;
 	int tid;
 	time_t start_time = -1;
 	double total_response_time;
